requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-telegram-bot[job-queue]>=20.0
sqlalchemy[asyncio]
asyncpg
python-dotenv
cryptography
tenacity>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
PyJWT==2.9.0
